# Conflicts:
#	src/main/java/controller/ReviewServlet.java
#	src/main/java/controller/UserServlet.java
#	src/main/java/controller/VerifyServlet.java
#	src/main/java/controller/WishListServlet.java
#	src/main/java/dao/AddressDao.java
#	src/main/java/dao/UserDao.java
#	src/main/java/dao/WishListDao.java
#	src/main/java/service/AddressService.java
#	src/main/java/service/UserService.java
#	src/main/java/service/WishListService.java
#	src/main/java/util/AuthFilter.java

+++ src/main/java/model/WishListRequest.java
@@ -8,6 +8,6 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @AllArgsConstructor
 public class WishListRequest {
-    int userId;
-    int bookId;
+    Long userId;
+    Long bookId;
 }
